Remove `controller_system_test.go`
* Keep `createDTMockClient` helper function for now, it's still used in the ActiveGate test
